Rename tsconfig paths for more clarity
Change-type: patch
