+++ lib/context.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '~/tests';
+import { expect } from '~/test-utils';
 import { Context } from './context';
 
 describe('Context', () => {
