+++ lib/diff.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '~/tests';
+import { expect } from '~/test-utils';
 import { DELETED } from './target';
 import { Diff } from './diff';
