+++ lib/testing.spec.ts
@@ -1,4 +1,4 @@
-import { expect } from '~/tests';
+import { expect } from '~/test-utils';
 import { plan } from './testing';
 
 describe('Testing', () => {
