+++ tests/orchestrator/tasks.ts
@@ -1,8 +1,8 @@
 import * as Docker from 'dockerode';
 import * as tar from 'tar-stream';
 
-import { DELETED, Disposer, Initializer, Target, Task } from '~/lib';
-import { console } from '~/tests';
+import { DELETED, Disposer, Initializer, Target, Task } from 'mahler';
+import { console } from '~/test-utils';
 import { App, Device, Service, ServiceStatus } from './state';
 import { getContainerName, getImageName, getRegistryAndName } from './utils';
