+++ tsconfig.json
@@ -12,16 +12,13 @@
 		"declaration": true,
 		"skipLibCheck": true,
 		"paths": {
-			"~/lib": [
-				"lib/index.ts"
-			],
 			"mahler": [
 				"lib/index.ts"
 			],
 			"mahler/testing": [
 				"lib/testing/index.ts"
 			],
-			"~/tests": [
+			"~/test-utils": [
 				"tests/index.ts"
 			]
 		}
